requirements.txt
streamlit>=1.29.0
yt-dlp>=2024.8.6
opencv-python>=4.8.0
pillow>=10.0.0
requests>=2.31.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0